Seperate inference from deps - TODO: need to sort out metadata
